Java API export of CorePlatformApi
This API surface is not a proper superset of PublicApi. e.g. consscrypt
provides SSLEngines to public's stub jar but not to coreplatform's stub
jar.

This complexity is not included in API export. The complexity will be
handled by the build orchestrator during API assembly (possibly via some
hardcoding).

(Also updating string representation of system_server to match the one
in Soong)

Test: b test //build/bazel/rules/apis:api_domain_test_suite
Bug: 262014796
Change-Id: I60c80d6b4d716c47024072519c84a7ed200dbc95