LSE fix to allow different operation kinds in the same heap location
We had a crash in which a vector operation was replaced by a scalar
value. This CL fixes that and allows vector and non-vector operations
to reference the same heap location.

Note that we can still use the knowledge of these HeapLocations to
propagate values, but we do not replace vector loads with scalar values
or viceversa.

Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling the app in the bug
Bug: 243136064
Fixes: 243136064
Change-Id: I2cf7d647225e32c111d57140bbca7042a7424667