Re-land "[GWP-ASan] Remove thread clamping in tests.""
This reverts commit 05a65e1d3e5f99ce1ec94817440114a1dce50978.

Reason for revert: Re-land the patch.

Originally merged in aosp/2420479, reverted in aosp/2419660.

-------

[GWP-ASan] Remove thread clamping in tests.

It's better and easier for us to just have threads contend against each
other in the tests if it's more than the maximum supported number of
hardware threads available.

Specifically, the recoverable test fails on Android because the
GTEST_SKIP in a called function, and it only properly works from the
TEST_* harness function. Android tests run on cuttlefish, which can be a
single core with two hyperthreads.

Reviewed By: fmayer

Differential Revision: https://reviews.llvm.org/D143221

Change-Id: I0cde38e4f2248d24bef56b975af029136e8556bb