Cvd stop will try stop_cvd as well
For some operations in generic CvdCommandHandler, it might be
sufficient to always return a static bin string as the name of
the executable. However, for others like stop, it should be
decided at runtime.

The CL opens a door to accomodate the logic to determine the
name of the executable binary at runtime, and applies the
logic to cvd_stop.

Bug: 267491049
Test: HOME=<home dir> cvd stop for old and new branches
Change-Id: I7594174a7448281a4762b6c07952e2f205d394e8