Do not reset IpClientCallbacks mock at end of performDualStackProvisioning.
performDualStackProvisioning helper fun is used to do dual-stack
provisioning in IpClient integration test to avoid duplicate code,
we do clearInvocations(mCb) at the end of this func to make test
case: testNoFdLeaks pass, otherwise, the test will fail due to the
mCb.onQuit are invoked multiple times because testNoFdLeaks does a
loop to repeat dual-stack provisionings to see if any sock FD leaks.

However, for some test cases that need to verify if mCb has executed
the sepecific method after performDualStackProvisioning, that gives
the error: the method is wanted but never invoked due to
clearInvocations(mCb) which happens earlier. So move this reset call
to testNoFdLeaks, for example, shutdownAndRecreateIpClient should be
a better place to reset mocked mCb before shutting down the IpClient
instance.

Bug: 266256943
Test: atest NetworkStackIntegrationTests
Change-Id: I71c5cdc2f08218eedd4ebb904113152e0ad6342c