Improve the dropbox app crash checking logic
Previously when CSuite checks whether a dropbox entry contains crash info for an app, it only checks whether the package name is present in the entry. This occasionally yields false positives as sometimes when a system process crash it may dump the running process info into the dropbox entry, which contains the app process name. This CL improves the logic by first trying to extract the source process name from a dropbox entry. If the source is not determined, it then use an improved regex to check whether the entry contains the package name.

Test: unit test
Fix: 259435400
Change-Id: Ie6d06254231acc82e3072f99ba650acdb641ac7a