Add wrapper functions for HMAC_CTX_*
Create a rust wrapper for HMAC_CTX_* functions to enable the iterative
HMAC APIs (calling update() for segments of the input rather than
collecting the whole input and calling hmac().

This patch is carried in AOSP because hmac.rs is not available upstream
yet, and the HMAC_CTX_* functions are recommended only for boringssl -
in openssl3 these functions are deprecated and the EVP_MAC functions are
recommended instead.

Bug: 267375624
Test: m libopenssl
Change-Id: I3d77403dd8dd607c8b55bf2be499e08895a76c19