Autoupdate of 3 submodules
*system/libvintf updated to b0cc671f7841533060b1691515f8f1e2ab483eae
Implement VintfObject::getLatestMinLtsAtFcmVersion.

This function returns the MAX of minlts of FCM at
a given level.

Details can be found in the kernel/configs project.
For example, for Android 12, this returns 5.10.43,
and for Android 13, this returns 5.15.41.

Test: TH
Bug: 255995840

Change-Id: Ib83b8013f4b903da6c943f45c572a15d7b2be997

*test/vts-testcase/hal updated to 0abcda9622a9a6d730041435ef990a7d87ab1757
Expose GetFcmVersionFromApiLevel.

Expose the map for other tests to use.

Test: TH
Bug: 255995840
Change-Id: I05e11ea150749ab876d8142fa2c26f25ec002e8b

*test/vts-testcase/security updated to e422bc040fe1a6e9fe3bdefc92dc1f80afea671c
Check kernel version requirements for GRF devices.

... as required on the VSR documents. GRF devices must always
run a kernel that is higher than the one frozen for the
board API level.

The latest minlts frozen for a board API level can be found
under kernel/configs. For example, for Android 12, the latest one
in kernel/configs/s is 5.10; and
kernel/configs/s/android-5.10/android-base-conditional.xml specifies
also the minor version (5.10.43).

Test: TH
Bug: 255995840
Change-Id: I9e71186098cf7cca6f11ed75bc69f86710443b23