Rewrite part of the unwind_for_offline tool.
The tool had a bug that caused it to not actually copy all of the
elf files. Rather than fix just that bug, rewrite the code to make
it a bit easier to understand and to take out some of the copy-paste
code to figure out when maps are from same elf.

Test: Ran on a process and verified the output could be put in
Test: an offline test.
Change-Id: Iada1e28716a089a274b32d9fef2c3385c62ca35b