Correct the Robolectric test type
Atest recognizes the test as a `modern` Robolectric test, which can run
via Tradefed, by detecting the existence of `Android.bp` in the module
path; however, this is not true when the mk and bp files coexist and
the test was defined in the mk file.

aosp/2308586 already set a test suite, `robolectric-tests`, for all
`modern` Robolectric tests in Soong, which is much more accurate, so
that this change leverage this information to determine if a test is
a `modern` Robolectric test.

Test: atest --host atest_unittests
Bug: 268202685
Change-Id: I7a4323f76bb3dd474f54e6b116d5ebbe1b9e3995