MicrodroidBuilder: Support CPU topology config
Setting the number of CPUs is now deprecated in favour of selecting
a CPU topology configuration, which determines the number of CPUs as
well as other properties such as CPU affinity, capacity, etc.

Keep the legacy numCpus option (apiLevel==33) but add support for
cpuTopology for newer Android releases (apiLevel>=34). Also check
apiLevel==33 for cpuAffinity.

Bug: 266664564
Test: atest -p packages/modules/Virtualization:avf-presubmit
Change-Id: I751e36852414a530f7a001230ff0adaec426799f