Remove libcplusplus from linkerconfig_diff_test
This seemed to only be added because python 2 binaries
needed libc++, but python 3 binaries don't.

We're making a change in soong to remove libc++ as a
dependency of the python3 interpreter. When we do that,
linkerconfig_diff_test_host_tools starts failing because
it can't access libc++ from its sandbox anymore.

Bug: Bug: 259718110
Test: atest --host linkerconfig_diff_test
Change-Id: Ib4ad17b11251f2728eef7bb8e27f674b315e0a5c