Fix ART MTS exclusion filter for libartpalette JNI test cases.
The change in the JniInvocation test in https://r.android.com/2375894
made us start the runtime with JNI_CreateJavaVM, which calls
android::InitializeNativeLoader, which in turn makes us load a mixture
of system and ART APEX libs in the same linker namespace, just like the
problem described in b/247108425#comment1.

We therefore need to exclude that test as well from MTS. The test is
primarily needed in CTS for native API coverage, so this is not too
bad.

Test: m mts && mts-tradefed run commandAndExit mts-art
Bug: 265674740
Bug: 247108425
Change-Id: I5711688376e0c896344ddcd72c25bb25549fb239