Autoupdate of 7 submodules
*build/soong updated to 97ce82e30f969ae054d3cb9e9d7c2a22b8c6c92c
Merge "rustc-1.66.1 Build 9591243"
*external/rust/crates/base64 updated to ee071cee30d09d169c4bd88d8bd71fc104b57071
Fix an error in the documentation strings

This file explicitly turns on this warning meaning that it can't be
overridden by command line lint arguments.  The rustdoc from version
1.66.1 will fail without this patch.

Test: m rustdoc
Bug: 263153841
Change-Id: Idcf3779cbd46300691232302bba10c46143a2dbc

*external/rust/crates/untrusted updated to b6a60db48458903531c1e4fa91ea8553e72bd535
Update crate source to work with Rust 1.66.1

This CL changes the name of a lint in an annotation to allow rustdoc to
successfully process the crate.

Test: RUST_PREBUILT_VERSION="1.66.1" m rustdoc
Bug: 263153841
Change-Id: If213ca89d296bb06d3bdeadb18b2c5b31ff36a2a

*packages/modules/Bluetooth updated to 088afbc340b55dca565d322e013729ac511b3b09
Merge "Update source for Rust 1.66.1"
*packages/modules/Virtualization updated to a096fed6f28910e848801ca9950e2e09d9545891
Merge "Update source for Rust 1.66.1"
*prebuilts/rust updated to 32fe2acd85cc0dd0e88b779f3216e11260c93147
rustc-1.66.1 Build 9591243

Bug: https://issuetracker.google.com/issues/263153841
Test: m rust
Change-Id: Id0f0df3048271c8672966587a392a1da0dcb66b2

*tools/security updated to a702306bf312411fb9d2db119695f69bbd57f967
Update source for Rust 1.66.1

This CL removes a feature flag for stabilized features.

Bug: 263153841
Test: m rust
Change-Id: I8439a045282d0208064189daa492c8d1ead7c491