tst_device: Add retry loop to wait for loop device to be available
There is some delay between /dev/loop-control
returning and the loop device actually being
created and accessible.  tst_device should check
the return value when it stats the loop device
and retry if it cannot access it yet.

We should determine if the delay is expected
behavior and that we are not masking a real
issue.

Bug: 148978487
Bug: 266990543
Test: atest vts_ltp_test_x86_64:syscalls.ioctl_loop04_64bit#syscalls.ioctl_loop04_64bit -- --abi x86_64
Change-Id: Ic6ab7532cd8e1bd01cf3b845b4a53c351f3396f9
Signed-off-by: Edward Liaw <edliaw@google.com>