Change order of HAL Data/Control packet definition
The UCI HAL Data packet has a constraint for the message_type field, while
the UCI HAL Control packet deson't (because it covers multiple values). This
leads to parsing errors for the UCI HAL data packet.

The workaround for this issue is to switch the order of these
definitions. This will result in packet bytes first being attempted to
be parsed as a UCI HAL Data packet, and then as a Control packet.

The code was tested by injecting Data packets in the UciManager (on a
HAL Open), and verified that the UWBSessionManager.onDataReceived() is
called. Tested for both unfragmented and fragmented data Rx packets
(https://paste.googleplex.com/5893713451548672).

Bug: 266016401, 269175546
Test: atest ServiceUwbTests FrameworkUwbTests
Change-Id: I887c9c9f2918f62e0260a3d3b79b225e5591a131