Relax CanMoveClinitCheck checks
In PrepareForRegisterAllocation, we can move the
responsibility from the clinit check to another instruction
e.g. NewInstance. We were checking that there weren't
throwing instructions between them, but they can exist.

Relax that check, and just don't move the responsibility
if such instructions exist in the middle.

Note that even when we can't move the responsibility to the
NewInstance, we can remove the ClinitCheck by moving said
responsibility to its LoadClass.

Bug: 268486612
Fixes: 268486612
Test: art/test/testrunner/testrunner.py --host --64 --optimizing -b
Test: dex2oat compiling the app in the bug
Change-Id: I0740a30a466ed0c7d5e2b069e63307ee706fd604