Use LD_LIBRARY_PATH instead of -Wl,-rpath.
Note: This affects both build.sh (deprecated) and Kleaf.

Drop the runtime library search path from ldflags so that
the resulting artifacts does not contain
the absolute path to the repository. This makes the artifact
more reproducible regardless of where the source code is.

If one needs to use a host binary outside of Kbuild, they
must provide LD_LIBRARY_PATH manually.

Test: TH
Bug: 268450910
Change-Id: I35ee44e6a9f62df969b14acd50375d45604d272d