Fix writeTypedList for min sdk < 23
The code written by GenerateTypedListHelper requires
Parcel#writeTypedObject, but that was only added in API level 23. So
if we need to work below that, instead use our (existing) helper for
writeTypedObject.

Also exercise a typed list in the tests. Note that this actually
passes both with and without the code generation change - it would
only fail if I were to test it on a pre-23 device, which would be
tricky. But we have evidence from Roblectric (see the bug) that it was
failing, and I've verified that it now works.

Bug: 269466052
Test: atest aidl_integration_test aidl_unittests
Change-Id: I83c9546a23865e3475cd93fb0e7a5bbbefb24a9a