Test: Add a DoH test for RESET_STREAM
In current DoH implementation, after receiving a RESET_STREAM
frame, DnsResolver is notified of a h3::Event::Reset with an
error when it is receiving the DoH response.

If there is any data received from h3::Event::Headers or
h3::Event::Data data, the data will be discarded.

The lookup thread that is waiting for the DoH answer will be
notified of DoH query failure.

Bug: 264966204
Test: atest
Change-Id: I6e63f1c6913c4ea2d5a2f205374adfd91733576f