Link libcrypto and libssl statically into libjavacrypto for host.
This avoids version skew between the libs in the
conscrypt-module-test-exports prebuilts, where an old prebuilt
libjavacrypto.so may otherwise load a too new libcrypto-host.so or
libssl-host.so built from source on master-art.

Also remove a stale TODO pointing to a resolved ticket. The current
plan to avoid a libjavacrypto prebuilt (as well as other prebuilts) is
go/multitree-design.

Test: Create and drop new prebuilts into a master-art tree, then:
      $ art/tools/buildbot-build.sh --host
      $ readelf -d out/host/linux-x86/lib64/libjavacrypto.so
      Check that shared library dependencies only include NDK libs.
      $ art/tools/run-libcore-tests.sh --mode=host -j48
Bug: 270126274
Change-Id: Iffb83f7ea123b588e3898e2ca437ad81d84b2f61