Clean up installed mainline modules for mainline test mapping suite.
This is a short term solution to fix flakiness and SLO in test mapping
suite. The root cause is the path of a mainline module after rollbacking
doesn't start with /system/, which impacting the optimization and
introduces the flakiness and breal SLO.

The change would be re-verted after chenzhu@ confirming with framework
team about why rollbacking app is not working.

Bug: 262626794
Test: unittest
Change-Id: I477fa342d71d3a196ebe46d12ff1a0d6f303da27