Insert missing type parameter list for psi based constructor item
Currently, Metalava omits type parameter list for constructor item when
generating API file. The discrepancy in the behavior between the method
item and the constructor item leads to unnecessary computation to search
for generics when generating stubs from the API text file, as seen in
aosp/2428997. This change modifies such behavior that the constructor
item and the method item is more closely aligned.

Test: ./gradlew
Bug: 269135125
Change-Id: Ib9950aad276259b00bc9bb2f7605cddbe99a0e4b