Autoupdate of 2 submodules
*device/generic/vulkan-cereal updated to 5b626d3497507b29ed4b81fca70fc379e62f6e41
Merge changes from topic "gfxstream-codecs"

* changes:
  gl-host: add api to get low level context
  osx-codecs: re-enable VideoToolBox decoder on OSX

*hardware/google/aemu updated to bdf64d7385cae203e3712835594883e6331f0171
osx-codecs: re-enable VideoToolBox decoder on OSX

1, As video tool box does not re-order frames, a buffer
of size 8 is used to sort them based on pts and then returns
to guest when buffer is full; this behavior depends on
aosp/1621650.
and the codec on osx should only be enabled on api31 with that
cl in the guest. otherwise, the decoded frames can be out
of order. for api30, this should be fine as omx uses real pts
instead of work index.

2, extract the w/h from sps/pps; avoided recreating decoder session
if the w/h do not change from previous sps/pps specification.

3, avoided scanning the input for next nalu when the first frame is
not SEI; when it is SEI, discard it and find the next nalu (should be
IDR)

4, added getLowLevelContext to EglDisplay that has non-null value only
in OSX; this context is used to create textures out of decoded frames.

5, performance: decoding into textures have good performance on M1
and on integrated GPU; not much on discrete GPU. It is suspected that
the memory transfer between discrete GPU and system memory is the
bottole neck.

for now, VTB is only enabled by ANDROID_EMU_MEDIA_DECODER_VTB=1
and it will be turned on when CTS passes fully on R and S regarding
media related tests using VTB.

Test:
decoding 4k buck bunny video,

Change-Id: I33f689340e60b949e33bdd0224f55049794cf8a6