audio: Remove InputStreamTest#getCapturePosition test
The test was created based on assumptions and the current
HAL behavior. There is no actual requirement in the HIDL
interface for a stream to have a zero capture position
prior to start. In fact, some streams may be created by
the HAL "pre-loaded" with data if audio has already been
captured by the DSP and needs to be passed to the framework.

Bug: 270478574
Test: atest VtsHalAudioV7_1TargetTest
Change-Id: Ied6a71b73bff403365dd85fe6a212cfa4a5b284d
(cherry picked from commit afe90852d4a9216e4468f7c5917047b43b44a877)