syscall01: use 32bit syscalls if available
For 32-bit applications, the getuid/getgid syscalls return 16-bit ids,
and the getuid32 and getgid32 syscalls return 32-bit ids.  When
CONFIG_UID16 is disabled in the kernel, getuid/getgid (16-bit UIDs) are
no longer available.  Thus this test will fail when compiled as 32-bit
and with CONFIG_UID16 disabled.  For 64-bit applications, this is not an
issue because getuid/getgid return 32-bit ids and getuid32/getgid32 are
not defined.

The fix for this is to use getuid32/getgid32 if they are available to
match the behavior of glibc.

Bug: 266732373
Test: vts_ltp_test_x86:syscalls.syscall01_32bit#syscalls.syscall01_32bit -- --abi x86
Change-Id: Ice03210c411b5df687c10687e13a3afc43b16bb0
Signed-off-by: Edward Liaw <edliaw@google.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
(cherry picked from commit a5d15b796362fd2731fa0290d48c096d370a040d)