ui: Remove state from controller/globals.ts
Previously the code under controller ran in a separate Worker.
This necessitated a bunch of complexity:
- The controller maintained a separate copy of the state
- Which it updated using 'patches' generated by applying actions
  on the main thread
- These patches had to be postMessage'd from the main thread to
  the controller thread

Since the controller code now runs on the main thread much of this
can be removed. Specifically:
- We no longer have to use a `Channel` and `postMessage` to trigger
  the controllers to run.
- The Controller code no longer needs to maintain a separate
  state object - instead all controller code can look at the
  'frontend' state object.
- A long unused `errorReportingChannel` can also be removed.

This change is very almost a noop. There is one specific behaviour
change I foresee. Prior to this change in the rare case where
controller code dispatches an action that action did not immediately
effect the state as seen by the controller - after this CL it does.

The continues the plan to:
- Remove unnecessary complexity of the controller/frontend distinction
- Remove controller/globals.ts entirely

Change-Id: I8ed10e1fcbebbdd6aa3abed45f64e294700c703c