Set adb shell screenrecord time limit to maximum
This change sets the time limit for the adb shell screenrecord command to the maximum value of 600 seconds (10 minutes). Previously, the command did not have a time limit set and defaulted to a 3 minute timeout. This caused issues when running crawl tests with a 10 minute timeout, as the screen recordings would only be 3 minutes long. With this change, the command is set to use the maximum time limit, allowing most tests to have full-length screen recordings. If a test runs for more than 10 minutes, the recording will still stop at 10 minutes, and it will be up to a future code change to address the issue by using multiple screenrecord calls.

Fix: 270986952
Test: forrest
Change-Id: Ic32e6a4456734972ac13b113039482ff0617cb60