Only tracing listeners need method entry for on stack methods
We need to report method entry events for on stack methods when enabling
tracing. This is required to match method exit events (that will be
reported when method exits) to method entry events. This is only needed
when starting method tracing and not required for jvmti listeners or
even tracing listeners once method tracing has already started. Earlier
we used to check for this by checking if there are method entry / exit
listeners and if interpreter stubs are installed. This used to work
because we forced interpreter whenever method entry / exit listeners are
installed. This is no longer the case and hence explicitly call
method enter events when starting method tracing.

Bug: 270380293
Test: art/test.py
Change-Id: I91b2aa473f2a2e5753fdac9d6f5ffcaed8a1749a