Autoupdate of 2 submodules
*frameworks/base updated to eaff057f6f057f1e8250267b1528a524a1519ec9
Merge "Pipe data from onWakeupEvent to batterystats"
*packages/modules/Connectivity updated to 1e31203fccbd939b6ec6cd75b969d9b25eb3d6bb
Add network handle to the nflog prefix for wakeups

This is useful in the system server for better attribution of wakeup
events.
The logic for encoding the data in the prefix is still maintained ad-hoc
and should be standardized in the future.

Test: Builds, boots. Manually triggering a wifi wakeup doesn't result
in anything unexpected.
Test: atest FrameworksNetTests

Bug: 265742148
Change-Id: I7ef8a606459bb9f0630f629c01d3b34643ae5319