Improve gcov kernel file collection performance
The approach GcovKernelCodeCoverageCollector used to collect gcov data
files out of debugfs and into a tar file was taken directly from the
suggested `gather_on_device.sh` script in the Linux documentation. This
approach used a combination of `find` and `cat` to locate all *.gcgo and
*.gcno files, `cat` each files contents into a temporary file and then
finally using `tar` on these temporary files to create the archive. This
approach was very slow but needed as the official gcov documentation
stated:
"Due to the way seq_file works, some tools such as `cp` or `tar` may not
correctly copy files from sysfs."

This is presumably because at the time both `tar` and `cp` early out of
their operations if the file is reported to be empty. All debugfs files
appear empty. This, however, changed for `cp` with the following change:
https://github.com/coreutils/coreutils/commit/4b04a0c3b792d27909670a81d21f2a3b3e0ea563

/* copy_file_range incorrectly returns 0 when reading from
   the proc file system on the Linux kernel through at
   least 5.6.19 (2020), so fall back on 'read' if the
   input file seems empty.  */

Therefore, we can now use a straight copy command of the folder instead
of the `find` + `cat` approach. This gives a significant performance
improvement to data collection off of device. Here are the before and
after times taken from CF x86_64:
 * `find` + `cat` + `tar`: ~155 seconds
 * `cp` + `tar`:             ~7 seconds

Test: GcovKernelCodeCoverageCollectorTest
Bug: 271174470
Change-Id: I96aba4ab2e92343e459eb407c51ad935c21aa541
Signed-off-by: Joe Fradley <joefradley@google.com>