Uniquify generated source filenames
Soong uses a distinct generated source path for each "gensrcs" rule in
the Android.bp file. However when trying out the autogenerated Bazel
conversion, the Bazel "gensrcs" rule uses a per-package path for
generated code. In other words, all "gensrcs" in the same BUILD.bazel
file use the same path.

This led to an error with the two generated "wayland-protocol.c" files.

The simple fix is just to use distinct file names for the two cases. The
file names do not otherwise matter, as the code is compiled into a
static library, which already has a distinct name for each case.

Note: external/wayland is not currently on the Soong allowlist for
Bazel, but this gets it ready for when it will be built that way.

Bug: 144185560
Test: b build //external/wayland:libwayland_client
Test: b build //external/wayland:libwayland_client_static //external/wayland:libwayland_server --config=android

Change-Id: I142cf0187a421b6b251996c05e4dea4cc3163ecd