Fix device death tests.
The death message is usually reported using LOG(FATAL) or
CHECK().  In device tests, these messages are reported to logd
instead of standard error and cannot be seen by GTest.  Add
a custom test init function which adds a logger that repeats
the message to stderr if we are in a death test child process.

Bug: 270535879
Test: presubmit, ran the test from the bug
Change-Id: Ibd23031260d3e75f634af6d434bbf010b277803b