bump_sdk.sh: list modules with new APIs
Explicitly list modules with new APIs in the git commit message. This
information is already present in the command line, also included in the
commit message, but this change makes the information more prominent.

Also use escape characters like '\n' when building the git commit
message string, and expand these at the very end. This allows the use of

  message+="foo\n"

instead of the clunkier

  message+="foo
  "

Also fix broken use of variable `bug_text`.

Bug: 271274016
Test: manual
Change-Id: I07b58a29c7414129a2f11c30d05c9b8d349c97e3