Make outout of cppbor::parse safely mutable
The parser makes use of "incomplete" object types that have a set size.
This means they cannot be updated after they are parsed, leading to
subtle and confusing behavior if one tries to mutate a parsed object.

For instance, if an array is parsed, then the caller tries to append
the array, the object ends up in a bad state.

Fix this by adding a "finalize" method to the incomplete types,
allowing them to be used while parsing but converted to safely mutable
cbor types once the parsing is complete.

Test: cppbor_host_test_external
Change-Id: I04e4f06ec10777aa49912520558583947af52dd2