kleaf: hermetic cc toolchain for android binaries.
Add support for cross-compiling Android libraries
and binaries.

cc_binary / cc_library are built with the target platform.
By default, the target platform is not set to Android. The target
platform may be set to Android using --config=android_arm64,
--config=android_x86_64, or --config=android_riscv64.

When one of these --config=android_* are set, --config=hermetic_cc
is implied so that only the hermetic CC toolchain is used. Host
CC toolchain is not considered.

One may use target_compatible_with to restrict whether a cc_binary
or cc_library works with the host or the device.

Test: bazel build //build/kernel/kleaf/tests/cc_testing:android_tests --config=android_arm64
Test: bazel build //build/kernel/kleaf/tests/cc_testing:android_tests --config=android_x86_64
Test: bazel build //build/kernel/kleaf/tests/cc_testing:android_tests --config=android_riscv64
  Note: riscv64 does not work yet due to the lack of NDK_TRIPLE. This
  will be fixed in follow up CLs.

Bug: 228238975

Include-Buildifier-Warnings: enabled

Change-Id: If15614ec5ffb21799f23315c586b62bf33d1fc7f