Construct a fake DEX2OATBOOTCLASSPATH for devices that don't have it.
This fixes the ParsedOptionsTest failure, which is another breakage
caused by aosp/2458432.

That test focuses on ParsedOptions and doesn't need a real
DEX2OATBOOTCLASSPATH or a real BOOTCLASSPATH. Real values can't be used
by the runtime in chroot anyway. However, the runtime needs them when
inferring the boot image location, while DEX2OATBOOTCLASSPATH doesn't
exist on old Android platforms. Therefore, we just pass a fake one to
make it happy.

Bug: 269230245
Test: art/tools/run-gtests.sh /apex/com.android.art/bin/art/x86_64/art_runtime_tests
Change-Id: Iaa047903c877d1640029f1166aa341353ceff24b