Remove SoongConfig.bp as it is no longer used
It is longer necessary to use `soong_config_module_type` to create a
custom module type in order to specify the Soong config variable to
use to control the `prefer` property. All existing usages of the module
types defined in this file have been replaced with usages of the
`use_source_config_var` property so this file is no longer needed.

Test: m nothing
Change-Id: I94569eac3c33dba2b4d77df7a6be826ced836784