Compile SdkSandbox against framework-sdksandbox.impl
SdkSandbox uses some private classes of framework-sdksandbox, so it
must compile against the implementation library, not stubs. This is
consistent with the comment that was already there.

I'm not sure why this wasn't causing a problem when compiled into an
apex, but with an upcoming change to make apex's dependencies
independently compilable, the issue was exposed.

This was also already done internally at ag/19378654

Bug: 254205429
Test: m SdkSandbox with aosp/2462194
Change-Id: I03f145f9bc0eb7f6868b5efaec6ea3afab992e3d
Merged-In: Icf0c3fc6e0560bbbef112034b0e0888a1b08c1f8