GKIComplianceTest: Account for running under a 32-bit userspace process
The GKI compliance tests assume that they are running under a 64-bit
process by only testing devices when uname reports "aarch64" as the
machine name. Uname returns "aarch64" if the process that invokes
uname is a 64-bit userspace process on a 64-bit ARM kernel.

However, the process can also be a 32-bit userspace process, in which
case uname will report "armv8l" as the machine name when using the
same 64-bit kernel. Thus, add "armv8l" as part of the list of machine
names that can be tested for GKI compliance.

Bug: 271889158
Test: atest vts_gki_compliance_test
Change-Id: I7ffc633df4c2499378301a9ce578b94b99eb86ed
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>