Fix rollback detection
After OTA finishes, update_engine writes boot id to /data/ to indicate
that a slot switch is scheduled. This is called a slot switch indicator.

After device reboots, update_engine uses value of the slot switch
indicator to determine if a slot switch was attempted for last reboot.
If slot switch was attempted but device is still on the old slot,
update_engine declares that a rollback has happened.

However, the slot switch indactor is only cleared when next OTA starts,
so if the device was rebooted during merge(or before next OTA starts),
update_engine would mistakenly think that a slot switch was attempted.

Test: reboot device during merge, make sure UE did not detect a rollback
Change-Id: Ia557cc14e4f217813fdcb7896b49131898e8bfd3