Remove obsolete code (following handle scope removal in generic JNI).
On arm64, the code that saved FP in generic JNI trampoline is no longer
necessary since https://r.android.com/98108 (note that the frame pointer
is already saved in X28, and FP was previously needed in addition to
X28).

The `BuildGenericJniFrameVisitor::FillJniCall` class is obsolete since
https://r.android.com/1582970 that removed handle scope and can be
replaced with its superclass `FillNativeCall`.

Test: treehugger
Change-Id: I83ba57795a088d32d2b14dd9d2e00f1755f26c35