Autoupdate of 2 submodules
*frameworks/support updated to 5799fcd16ba2f59e05a95fff354b505b44074326
Merge "Adds support for headless date pickers" into androidx-main
*golden updated to 4623d118bc2f9e4a4ba1121e6dcc381dc2000ec4
Updates the date pickers goldens

Updates to the date pickers Scuba goldens to reflect the changes in the
layout of the internals of the components.
Most of the changes here reflect the inclusion of the divider between
the header and the picker's calendar into the header's height.

Change-Id: I573b35584b61e8f3140d5d839ace6e3a43769a81