Treble VINTF test: Allow vendor extensions for passthrough HALs.
The change cc24ca47869880b25f33288469cc2ea3674bbaf7
"Treble VINTF test: parameterize on HAL instances."

  https://android.googlesource.com/platform/test/vts-testcase/hal/+/cc24ca47869880b25f33288469cc2ea3674bbaf7

accidentally introduced a requirement on passhthrough
vendor extension HALs that the dynamic library (.so)
with the name of the vendor extension must exist. For example,
if vendor.foo.display.mapper extends android.hardware.graphics.mapper,
it was erroneously required that vendor.foo.display.mapper@3.0-impl.so
exists. This was because when the test cannot find
vendor.foo.display.mapper@3.0-impl.so, it skips calling interfaceChain
on it, thus it cannot infer that it extends a HAL that is in
the passthrough allowlist.

Remove this incorrectly imposed requirement by actually searching for
all passthrough HALs on the device and look for the implementation
that also implements vendor.foo.display.mapper@3.0. Then, look up
intefaceChain() on the implementation to match against the allowlist.

Bug: 261765512
Test: vts_treble_vintf_vendor_test --gtest_filter=*vendor* on a pixel
  device

Change-Id: I06bc822e3af82ad5c7ca53207a10175610779b51