Fix DeviceOwnerTargetPreparer for headless system user
Removes functionality that assumes the primary user is not 0, as
in practice they are the same.

Also removes device admin from main user for HSUM devices to allow
the package to be uninstalled.

And implements the uninstall logic so that tests don't have to
manually add an final uninstall step.

Bug: 267492062
Bug: 268271623

Test: tradefed > run host --class
    com.android.tradefed.targetprep.DeviceOwnerTargetPreparerTest

Change-Id: I18964355c92f26cc91c917e726e89eb20ef5dedd