[scudo] Add a method to force release everything.
The force flag to releaseToOSMaybe does not release everything
since it is an expensive operation. Modify the release flag to
have three states: normal, force, forceall. Force behaves the same
as setting Force to true from before this change. Forceall will
release everything regardless of how much time it takes, or
how much there is to release.

In addition, add a new mallopt that will call the release function
with the forceall flag set.

Reviewed By: Chia-hungDuan

Differential Revision: https://reviews.llvm.org/D146106

GitOrigin-RevId: 8aaefa92436cf7f66e107795d8513db1285952ac
Change-Id: I92fe9b98b4dd9ca584722d842cfc64cc87b42e00