Autoupdate of 2 submodules
*external/cronet updated to 4ba5a55364fe402781ef015011f57c834f1f5ba3
Use the version string without sha1 prefix for public purposes in AOSP.

Imports to AOPS should be limited to stable builds which have a well
defined version so the SHA isn't necessary (this should be the case for
other channels too in case we need to switch at some later date).
Moreover, the filling in the SHA would require additional infra
support (b/271846308).

Test: m
Change-Id: I334dcb4d11e31e52b5cca189ad9ba9ec531096f6

*packages/modules/Connectivity updated to 020b31904a52d974cf2da77289c815c3600e157b
Merge "Add a test for the version string presence."