Don't immediately abort for large maps and arrays
When the parser finds a map or array it will try and reserve space for
all of the elements. However, this leads to allocation failures when the
array or map claims to have a large number of items which, in turn,
aborts the process.

By removing the memory reservation and only allocating as the items are
parsed, the memory allocation is limited to a multiple of the input size
which reduces the chance of an abort. This may result in more
allocations and copying of data but there aren't any known cases where
this performance is important.

Bug: 273450455
Test: atest cppbor_host_test_external
Change-Id: Ieede0ee3651eb111bad6f1930c77dc510b6defeb