Add IFA_F_MANAGETEMPADDR and IFA_F_NOPREFIXROUTE flags in ConstantsShim.
These two flags are assigned when installing an IPv6 address which is
generated from delegated prefix, the constants were instroduced since
API 34 in OsConstants.

The IFA_F_MANAGETEMPADDR flag tells kernel to create an extra temporary
address associated to this IPv6 address and manage it on behalf of user
space, e.g. the temporary address lifetime should be also extened when
extending this address; or delete them together when receiving an RA
with 0 PIO preferred/valid lifetime.

The IFA_F_NOPREFIXROUTE flag tells kernel to suppress the creation of
route for delegated prefix, to prevent the traffic loop between the host
and upstream delegated router.

Bug: 260934173
Test: m
Change-Id: I042818f94978fdced2b821c849ab2645a5b723dc