comments for not differentiating AIDL
The same AIDL files must generate the same code
in different environments. There are a few bugs
referenced in the code, so not mentioning here.

This is important for:
- avoiding ODR issues (mixing 'aidl_interface' and
  'core' build system) or when core build system
  is used to compile the same AIDL file multiple times
  with different options (this was a requirement
  originally made by ctate@ and hackbod@ when they agreed
  to let us take the unify android idl project)
- when we want to switch an interface between stable and
  unstable, including for staging HAL interfaces

Noting in the code so that these requirements are
documented, but we may want to follow-up with more work
to prevent these languages from diverging.

Test: N/A
Change-Id: I203e5614988fcf9528ea2a6636f4b2eeedd003eb