ANDROID: make policy-inliner.sh stricter
Before, if an argument was missing, the corresponding bash variable
would be empty string and then the `cat` command would happily succeed
by outpuing just a new line.

I spent a lot of time debugging an issue that was caused by missing
syscalls in one of the policies because I forgot to add the new include
file to the build rule.

Test: m
Change-Id: I9575dbd62008cf3ee72bddfb213e18d04ff1e18e