Remove receivers from extension functions when listed as class methods in Java
I think this is the simplest way to resolve b/236889552, and the receivers aren't needed on *Kt class pages, since the receiver in Java should always be the class itself.

This doesn't change how functions are displayed in the extension function section, they will still have receivers there.

This isn't needed for properties because extension properties in Java always become accessor functions first.

Test: Added a test of extensions defined within a classlike, updated integration test outputs to remove receivers from Kt pages.
Bug: 236889552

Change-Id: I29f10ea2a7d38ea726b57a095cb9eb93e7e6a6da