Remove the "stem" field from a target definition
This CL modifies the target definition for the `log` crate by removing
the "stem" field.  Previously Soong was building static versions of this
library as `liblog.rlib`.  Unfortunately rustc would find `liblog.a`,
the C++ loging implementation, and assume it was the required static
library for the `log` crate and pass that into the linker, resuling in
linker confusion as expected symbols were missing.

The CleanSpec.mk file is needed to remove the intermediate directory for
the targets for this crate.  If both `liblog.dylib.so` and
`liblog_rust.dylib.so` are in the same directory the resulting
`keystore2` binary will crash when the device boots.

Test: Cherry-Pick ag/21614245 and run `m libukey2_jni_shared`
Change-Id: I547f7b9f8c9b3443b874f632a88cfa7e3f86270e