flatten directory structure
This removes many of the nested directories here
so that everything as at the top level. At this point,
everything is removed which can't be used.

The only directory which isn't flattened is
tools/vts-core-tradefed, and this is because
it has a config.mk file in it referenced by the core build
system. I think this configuration really could be placed
in the core build system.

There are a few tests here, and they should likely be
moved to be next to the components they test.

vts-core-tradefed can be deleted entirely, but there
are a few preparers which some tests depend on, and
some tests also depend on 'isReadWriteFile', so they
would need to be re-written.

Bug: 274790216
Test: N/A

Change-Id: I9d14d96cd14641a35237a16ba6db7d80a35f308b