Make the palette client tests independent of the io cgroup controller
CL aosp/2378990 migrates the io cgroup controller from the v1 to the v2
hierarchy on the master branch but not on the udc-dev branch. Because of
the migration of the io controller from the v1 to the v2 cgroup
hierarchy, testing whether cgroup support is available is not sufficient
to use an I/O profile. Change "NormalIoProfile" into "ProcessCapacityHigh"
to make these tests independent of the io cgroup controller.

This CL has been tested by running the following command against
Cuttlefish: atest -s 0.0.0.0:6520 art_standalone_libartpalette_tests
As one can see, no tests were skipped:

    Running Tests...

    x86_64 art_standalone_libartpalette_tests
    -----------------------------------------
    art_standalone_libartpalette_tests (6 Tests)
    [1/6] PaletteClientTest#SchedPriority: PASSED (4ms)
    [2/6] PaletteClientTest#Ashmem: PASSED (4ms)
    [3/6] PaletteClientTest#Trace: PASSED (12ms)
    [4/6] PaletteClientTest#SetTaskProfilesCpp: PASSED (5ms)
    [5/6] PaletteClientTest#SetTaskProfiles: PASSED (11ms)
    [6/6] PaletteClientJniTest#JniInvocation: PASSED (195ms)

Bug: 213617178
Bug: 274930420
Change-Id: I5d8ddb53702ce411b7a18901c3207adfefd8c12c
Signed-off-by: Bart Van Assche <bvanassche@google.com>