Autoupdate of 2 submodules
*test/vts updated to b1fe4ce88a204d1b7bdb9e0ad0b82a0b0fd5f949
move out TargetFileUtils

This is the only real thing in vts-core-tradefed-harness
that tests are using, so moving it to core tradefed.

Bug: 274790216
Test: atest FirmwareDtboVerification FirmwareBootHeaderVerification KernelSelinuxFileApiTest
Change-Id: I828ece56bcdd88e2e0f917e98917421474911c14

*tools/tradefederation/core updated to f83fcc571a8eda45fa04b3705edb4b7ab37d6b72
add TargetFileUtils to core tradefed

This is the last utility in VTS that is used by many tests
and prevents us from deleting vts-core-tradefed-harness.
The only other thing remaining is the kernel test runner.

I couldn't find an alternative for this functionality, so
I've moved it as-is, except I replaced use of EasyMock
with Mockito, as seems to be the style here.

The code could be simplified, but more important is avoiding
forking our test infra like this.

Bug: 274790216
Test: m tradefed-all && ./javatests/run_tradefed_tests.sh
Change-Id: Iad791e5abd889076920b761140d17f35c8b3d20e