riscv64: Let dex2oat run in verification mode (disable compilation).
There is no compiler support for RISC-V yet, but dex2oat can perform
verification and compile boot images and app images as though it was
running with compiler filter "verify". This allows us to re-enable
dexpreopt and speed up the first boot, as there is no need to verify
all the apps on device.

The change is done purely in ART by adding a workaround to turn off
compilation completely on RISC-V, as opposed to setting compiler filter
"verify" in the build system (e.g, we still keep compiler filter
"speed-profile" for boot images in RISC-V). This approach is chosen so
that fewer modifications are needed when adding compiler support.

Bug: b/271573990
Test: `lunch aosp_riscv64-userdebug && m`
      (with cherry-picked https://r.android.com/2511036)
Change-Id: If5dd798d043edc0c0ed321ad2ef5b841e68efe0f