Support multi-byte key usage with endianness workaround in VSR 14
The DICE key usage field is allowed to consist of multiple bytes in
little-endian order so accept that encoding.

For compatibility with ROMs that produce a multi-byte key usage with the
wrong endianess, also allow valid big-endian encodings of the field in
VSR 14 and assume the key can be used for verification. This is unlikely
to cause problems and allows support for more SoCs.

Bug: 273552826
Test: atest --host hwtrust_cli_tests hwtrust_tests libhwtrust_tests
Change-Id: I0a4b21fe3f91f9c3934373cfadebcebddfd4bb1f