Add support for overriding the fetch CVD path, which made to allow for more flexibility in specifying the fetch CVD path.
This change adds support for overriding the fetch CVD binary. If a new fetch CVD path is specified in the extra oxygen args, it will be added to the list with the "-override_fetch_cvd_path" flag. If the path is not specified, the build target/build id will be read from BuildInfo and used to construct the fetch_cvd path.

Bug: 273401628
Change-Id: I8cc2502b238e0d605c0a67ca650fdff6f84a0866
Test: presubmit / unit test