Regenerate the workspace when checksum file parsing fails
Atest Bazel mode analyzes an MD5 checksum proto file in each
invocation to determine whether the workspace should be generated.
This format of this file was recently changed to protobuf.

Atest crashes on local workspaces that still use the legacy MD5
checksum file format.

This change makes sure that we treat the file as non-existent and log
a warning when it can't be parsed.

Test: atest atest_unittests && m atest && atest-dev atest_unittests
Fix: 276498394
Change-Id: Ic88c5b3c6d64bc180e182007cf0feedc3e4e6ac5