Autoupdate of 2 submodules
*device/generic/goldfish-opengl updated to 21488dc2e3c7b10ca7a419a8d9f7c705bc627397
goldfish-opengl: guest memory

This adds guest memory support to gfxstream for non-shareable
Vulkan buffers.  It uses an experimental extension to the blob
API (F_CREATE_GUEST_HANDLE) to do this.  Once we have things
working on the target platform, the plan is indeed to upstream it.

This was tested on a Linux setup with udmabuf.  The standard tests:

./deqp-vk --deqp-case=dEQP-VK.memory.*

have a 95% pass rate.  The failures are related do the host kernel
not being able to create an udmabuf (i.e, VK wants buffers that are
too large).  This is a Linux kernel quirk and not related to
gfxstream.

There is the addition of a CPU timeline, so that gfxstream guest
can be sure that the buffer has been created before proceeding with
further GPU operations.

Overall, this shows how guest memory can work and how target
platforms can adopt via virtio-gpu rather vendor specific interfaces.

BUG=262595016
TEST=compile and run CTS

Change-Id: I3cd20406f13e3fd643378347db501264d4dddb45

*device/generic/vulkan-cereal updated to e02e4a2661dc681d72b7f1472308a4a5df81050c
Merge changes from topic "gfxstream-guest-mem"

* changes:
  vulkan-cereal: guest memory prototype
  vulkan-cereal: deferred resource creation host-side implementation
  vulkan-cereal: codegen for vkGetBlobGOOGLE and VkCreateBlobGOOGLE