Unique UDP encap sockets per IKE session
This patch changes the uniquness mapping of IKE sockets from being
unique on a per-Network basis to a per-IkeSession basis. This allows for
sockets to follow IKE sessions around, and creating a simpler 1:1
mapping of sessions to sockets. Additionally, this allows for kernel
migration without needing to update encapsulation sockets, which has
additional complexities around updating encapsulation types.

Bug: 179068694
Test: atest FrameworksIkeTests CtsIkeTestCases
Change-Id: I1759b9f1fcabf66dc309f44f7ab711ffcf92eed3