Python: prevent deleting existing venv directory
When `--venv-dir` is set, `mVenvDir` is used, but should not be
removed on cleanup since it's provided by the user. Since `mVenvDir` is
added to build info versioned files using `buildInfo.setFile` during
`installDeps`, the venv directory is then deleted on cleanup.

To prevent this deletion of the user provided venv (ie. when `--venv-dir`
is set), create a symlink to temporary path within the work folder
instead.

Change-Id: Iaf6a8dc3a4bac3531d3ec3e4c6e4b04262c5b08f