Autoupdate of 2 submodules
*frameworks/libs/net updated to bdb74024f5620880542e3eb6d8f5e9dd928cf9f4
Merge "Add a build rule for files used by mdns"
*packages/modules/Connectivity updated to 6e12ac60b8704d2188713a573fed69fc8db79fcb
Refactor SocketNetlinkMonitor to make it buildable with system sdk

NetlinkMonitor cannot be used when building with system sdk. This CL
refactored the SocketNetlinkMonitor into a separate class and use
factory method and droidstub to avoid referring NetlinkMonitor when
building with system sdk.

Test: m service-connectivity-mdns-standalone-build-test
Bug: 272392042
Change-Id: I2d97f6fe8a17febd155c81303d86cd558654ddf5