Autoupdate of 2 submodules
*hardware/interfaces updated to de49fca027e410db5d0890fcbdd6df8ac2ab1b4b
Merge "Document that the component version can be a string"
*tools/security updated to 264fa1fea5578173f4cc0a23a4e975b965867384
Allow the component version to be a string from VSR 14

After receiving feedback from partners and having the need for
compatibility, allow the component version to be a free-form string from
VSR 14. Previous versions will still require the version to be an
integer.

Bug: 273552826
Test: atest --host hwtrust_cli_tests hwtrust_tests libhwtrust_tests
Change-Id: Ie18a870b32dc5397b4ba0fa1a75d699b5d5ff286