Speed up updater.sh by removing lunch dependency.
Since we only need to build a single host module, we can skip a lot of
the typical build setup. This reduces the iteration time of
external_updater to ~5s on my machine.

To eliminate the need of the caller to have run lunch as well, the
dependency on ANDROID_BUILD_TOP was also removed in favor of using the
CWD. It doesn't appear that there are any cases where the CWD was not
valid.

Bug: None
Test: ./updater.sh check libxml2
Change-Id: Ibb7209a4046068475977daddd142f248079f7e68