Disable SSH agent authentication for on-VM testing scripts.
Add SSH option `-o IdentityAgent=none` to the scripts that setup the VM,
connect to it and run the tests on it. Also propagate it to rsync
commands via RSYNC_RSH, and to scp (all commands that work over SSH).

This is needed to avoid SSH agent interfering with the preferred
authentication method and running interactive hooks that require manual
user intervention while running the scripts.

Alternative workaround would be to set SSH_SOCK_AUTH= (or any invalid
value). Previously it was set to an invalid agent, that's why the test
ran without interference.

Bug: b/271573990
Test: run any ART test on a Linux RISC-V VM and observe that it passes
  without requiring any manual intervention from the user:

  lunch aosp_riscv64-userdebug

  export ART_TEST_SSH_USER=ubuntu
  export ART_TEST_SSH_HOST=localhost
  export ART_TEST_SSH_PORT=10001
  export ART_TEST_ON_VM=true

  . art/tools/buildbot-utils.sh
  art/tools/buildbot-build.sh --target

  # Create, boot and configure the VM.
  art/tools/buildbot-vm.sh create
  art/tools/buildbot-vm.sh boot
  art/tools/buildbot-vm.sh setup-ssh  # password: 'ubuntu'

  art/tools/buildbot-cleanup-device.sh
  art/tools/buildbot-setup-device.sh
  art/tools/buildbot-sync.sh

  art/test.py --target -r --no-prebuild --ndebug --no-image \
    --64 --interpreter 001-HelloWorld

Change-Id: I2b1d48857074ade75e841a3008486caacbd32479