Cache the relative path to product out for module-info.json
Atest Bazel mode caches the content of module-info.json with absolute
path. When users switch build target, a new module-info.json will be
generated in a new place, and the old module-info.json is still under
the old product out directory. But Atest Bazel mode will still check
the change on the old module-info.json to determine whether to
regenerate the workspace, which causes inconsistent between the
workspace and module-info.json in use.

Test: m atest && atest-dev atest_unittests
Fix: 274999420
Fix: 277656887
Change-Id: I35d974c739b2650e4de5a7c84ae82231d60d47f6