Merge tag 'android-s-beta-5' into android12-gsi
Android S Beta 5

* tag 'android-s-beta-5':
  rb5: Add support for Qcom Robotics RB5 development kit
  rb5: Add bootloader and firmware binaries
  Add the 'bdev_type' attribute to all block devices
  db845c.mk: Do not copy kernel modules to vendor image
  init.common.usb.rc: Do not hardcode sys.usb.controller property
  db845c: Enable metadata encryption
  db845c: Switch gralloc to use gralloc.minigbm_msm instead of gralloc_gbm
  db845c: Update PRODUCT_SHIPPING_API_LEVEL to 31 and FCM to 6
  db845c: Update PRODUCT_SHIPPING_API_LEVEL to 30 and FCM to 5
  db845c: Bump FCM level to 4
  db845c: Switch to using android13-5.10 kernel by default
  db845c: Set ro.sf.lcd_density=160
  dragonboard: fstab.common: Set metadata as first_stage_mount to resolve adb remount overlay persistence issues
  sepolicy: Rename rtc to wakeup label and add other missing sysfs_wakeup labels
  pixel3_mainline: device.mk: add lcd_density for Pixel3
  device-common: make lcd_density device-specific
  pixel3_mainline: device.mk: make fstab device-specific
  pixel3_mainline.mk: Update vendor-only modules
  suspend_blocker: Use Cuttlefish's workaround to avoid suspending the device

Background:
The android12-gsi branch is the correct branch that everyone should use for the Android12 version development, but it misses some changes like the PRODUCT_SHIPPING_API_LEVEL updates which are necessary for the Android12 Version, so the android12-gsi branch needs to be updated with some changes after it was created. And about the reason to merge the android-s-beta-5 branch here is that the Linaro LKFT project uses this branch since the setup for the Android12 version projects, and the android-s-beta-5 was the one branch close to but not too far from for the Android12 version development at the point it's selected, and since it's been tested by the LKFT projects, it's the best option to be used for the merge here.

Test: boot to homescreen on db845c
Change-Id: I788a536ce88e2a04a8a0c47ea110fcff35aa2202
Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>